+++ DSAObjectShieldAndParry.h
@@ -39,10 +39,10 @@
                    inCategory: (NSString *) category
                 inSubCategory: (NSString *) subCategory
              inSubSubCategory: (NSString *) subSubCategory
-                   withWeight: (NSInteger) weight
+                   withWeight: (float) weight
                     withPrice: (float) price
                    withLength: (float) length
-                  withPenalty: (NSInteger) penalty
+                  withPenalty: (float) penalty
         withShieldAttackPower: (NSInteger) shieldAttackPower
          withShieldParryValue: (NSInteger) shieldParryValue
                 withHitPoints: (NSArray *) hitPoints
